requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
asyncpg==0.28.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-dotenv==1.0.0
structlog==23.2.0
aiohttp==3.9.1
prometheus-client==0.18.0
prometheus-fastapi-instrumentator==6.1.0
requests==2.31.0
PyYAML==6.0.1
SQLAlchemy
starlette==0.27.0
